Migrate from Hyperapp to Vue
Hyperapp is in a weird state between v1 and v2. Currently v2 is
in pre-release and there's next to no documentation. I attempted to
describe the clock web worker as an effect using hyperapp-fx in v1
and using the new effect API in v2 and neither worked.

Vuex cleanly handled the clock web worker as an effect, and Vue
supports JSX just fine. Also Browserify continued to get on my nerves
so I migrated over to Webpack.