Bulk WIP
It's been so long since I've worked on this I'm not sure what the state
of everything is. Trying to unload all of my personal stuff off my work
laptop, so I'm just bundling everything as-is and will try to break this
commit up into meaningful commits if I ever pick this back up again.